# archive/v3_cleanup_backup/cbs-platform/platform/api-schema/openapi-v2.yml
openapi: 3.0.3
info:
  title: CBS_PYTHON V2.0 - Core Banking System API
  description: |
    Advanced microservices-based Core Banking System with comprehensive financial services.
    
    ## Features
    - Microservices architecture with domain-driven design
    - OAuth2 and API key authentication
    - Real-time event streaming
    - Advanced rate limiting and throttling
    - GraphQL support for complex queries
    - Webhook notifications
    - Comprehensive audit logging
    - Multi-tenancy support
    
    ## Security
    - JWT-based authentication
    - Role-based access control (RBAC)
    - API key management
    - Request/response encryption
    - PCI DSS compliance
    
  version: 2.0.0
  contact:
    name: CBS Engineering Team
    url: https://cbs-platform.com/support
    email: api-support@cbs-platform.com
  license:
    name: Proprietary
    url: https://cbs-platform.com/license
servers:
  - url: https://api.cbs-platform.com/v2
    description: Production API
  - url: https://staging-api.cbs-platform.com/v2
    description: Staging API
  - url: http://localhost:8000/v2
    description: Development API
security:
  - BearerAuth: []
  - ApiKeyAuth: []
  - OAuth2: []
paths:
...
  /customers:
    get:
      tags:
        - Customers
      summary: List customers
      description: Get paginated list of customers
      operationId: listCustomers
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - name: status
          in: query
          description: Filter by customer status
          schema:
            type: string
            enum: [active, inactive, suspended, pending]
        - name: search
          in: query
          description: Search customers by name, email, or phone
          schema:
            type: string
      responses:
        '200':
          description: Customer list retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerListResponse'

    post:
      tags:
        - Customers
      summary: Create customer
      description: Create a new customer account
      operationId: createCustomer
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '201':
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
  /customers/{customerId}:
    get:
      tags:
        - Customers
      summary: Get customer
      description: Get customer details by ID
      operationId: getCustomer
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CustomerIdParam'
      responses:
        '200':
          description: Customer details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Customers
      summary: Update customer
      description: Update customer information
      operationId: updateCustomer
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CustomerIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerRequest'
      responses:
        '200':
          description: Customer updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'

    delete:
      tags:
        - Customers
      summary: Delete customer
      description: Soft delete customer account
      operationId: deleteCustomer
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CustomerIdParam'
      responses:
        '200':
          description: Customer deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'

  # Account Management
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
    
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.cbs-platform.com/oauth/authorize
          tokenUrl: https://auth.cbs-platform.com/oauth/token
          scopes:
            read: Read access to resources
            write: Write access to resources
            admin: Administrative access
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortParam:
      name: sort
      in: query
      description: Sort field and direction (e.g., name:asc, created_at:desc)
      schema:
        type: string
        default: created_at:desc
    CustomerIdParam:
      name: customerId
      in: path
      required: true
      description: Customer unique identifier
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
        - request_id
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        request_id:
          type: string
          format: uuid
          description: Unique request identifier
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            validation_errors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
                  code:
                    type: string
    SuccessResponse:
      type: object
      required:
        - success
        - message
        - timestamp
        - request_id
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          description: Success message
        data:
          type: object
          description: Response data
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
    PaginationMeta:
      type: object
      required:
        - page
        - limit
        - total
        - total_pages
      properties:
        page:
          type: integer
          description: Current page number
        limit:
          type: integer
          description: Items per page
        total:
          type: integer
          description: Total number of items
        total_pages:
          type: integer
          description: Total number of pages
        has_next:
          type: boolean
          description: Whether there is a next page
        has_previous:
          type: boolean
          description: Whether there is a previous page

    # Health and System Schemas
...
    CreateCustomerRequest:
      type: object
      required:
        - first_name
        - last_name
        - email
        - phone
        - date_of_birth
        - address
      properties:
        first_name:
          type: string
          maxLength: 50
        last_name:
          type: string
          maxLength: 50
        email:
          type: string
          format: email
        phone:
          type: string
          pattern: '^\+?[1-9]\d{1,14}$'
        date_of_birth:
          type: string
          format: date
        address:
          $ref: '#/components/schemas/Address'
        kyc_documents:
          type: array
          items:
            $ref: '#/components/schemas/KYCDocument'
    UpdateCustomerRequest:
      type: object
      properties:
        first_name:
          type: string
          maxLength: 50
        last_name:
          type: string
          maxLength: 50
        email:
          type: string
          format: email
        phone:
          type: string
          pattern: '^\+?[1-9]\d{1,14}$'
        address:
          $ref: '#/components/schemas/Address'
    CustomerResponse:
      type: object
      required:
        - id
        - first_name
        - last_name
        - email
        - phone
        - status
        - created_at
      properties:
        id:
          type: string
          format: uuid
        customer_number:
          type: string
          description: Unique customer number
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        date_of_birth:
          type: string
          format: date
        status:
          type: string
          enum: [active, inactive, suspended, pending]
        kyc_status:
          type: string
          enum: [pending, verified, rejected]
        address:
          $ref: '#/components/schemas/Address'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CustomerListResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/CustomerResponse'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    Address:
      type: object
      required:
        - street
        - city
        - state
        - postal_code
        - country
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        postal_code:
          type: string
        country:
          type: string
          minLength: 2
          maxLength: 2
    KYCDocument:
      type: object
      required:
        - type
        - document_url
      properties:
        type:
          type: string
          enum: [passport, driver_license, national_id, utility_bill]
        document_url:
          type: string
          format: uri
        verification_status:
          type: string
          enum: [pending, verified, rejected]

    # Account Schemas
...